refactor(ndk): move JSON serialization funcs into own file
