+++ bugsnag-plugin-android-ndk/src/test/cpp/test_utils_serialize.c
@@ -1,6 +1,11 @@
-#include <featureflags.h>
-#include <greatest/greatest.h>
+#include <fcntl.h>
 #include <stdlib.h>
+#include <unistd.h>
+
+#include <greatest/greatest.h>
+#include <parson/parson.h>
+
+#include <featureflags.h>
 #include <utils/serializer.h>
 #include <utils/serializer/migrate.h>
